[bd-ykgx9] Fix remaining AR amount_cents i32→i64 in generated types (MaroonHarbor)
Regenerated AR openapi.json + client. Zero i32 amount_cents remaining.

+++ clients/ar/src/types_2.rs
@@ -104,7 +104,7 @@ pub struct UpdateCustomerRequest {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct UpdateInvoiceRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub amount_cents: Option<i32>,
+    pub amount_cents: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub due_at: Option<chrono::DateTime<chrono::Utc>>,
     #[serde(skip_serializing_if = "Option::is_none")]
